api-server-flask/requirements.txt
# Flask and Core Dependencies
Flask==2.1.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.1
flask-restx==0.5.1
Werkzeug==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
itsdangerous==2.1.2
click==8.1.3

# JWT and Security
PyJWT==2.10.1

# MySQL Database Dependencies
PyMySQL==1.0.2
mysqlclient==2.1.1

# Excel and CSV Processing Dependencies - REQUIRED for file uploads
openpyxl==3.1.2
xlrd==2.0.1
xlsxwriter==3.0.9
et-xmlfile==1.1.0

# Date and String Processing
python-dateutil==2.8.2
chardet==5.1.0
charset-normalizer==2.1.1

# Web Server
gunicorn==20.1.0

# Development and Testing
python-dotenv==0.19.2
requests==2.28.1

# Utilities
certifi==2022.12.7
idna==3.4
urllib3==1.26.13
six==1.16.0
setuptools==75.6.0
pandas==1.5.3
# Note: Removed all SQLAlchemy dependencies
# Removed: Flask-SQLAlchemy, SQLAlchemy, greenlet
# This application now uses direct MySQL connectivity via PyMySQL